run_kit/templates/providers/gemini/requirements.txt
# Google (Gemini) requirements
google-generativeai>=0.3.0